Add sort todos by owner
This adds the radio button for and ability to sort the todos by owner.  I kind of hate the code because it seems repetitive to me, but I am committing because it is passing the tests and that is a win.  this adds a function to the todo service that will sort the todos alphabetically by their owner, this function is then called by 'updateOwnerSort()` when the button is pushed.