Add final todo sort
This adds the final sorting method.  I think they are ugly but I am tired and need to go to bed so I am committing here.  Right now there are two methods being used for each sort attribute and I would like to reduce that a bit.  That will be my goal for tommorw.  But for now they at least do a thing.