Refactor sorting todos
This reduces some of the code that is being used to sort the todos.  Instead of having a separate update function for all of the attributes, this has one single sort function that takes in a sort field, and sorts based on that sort field.  This also updates the HTML to reflect these changes.